Otimize the code

+++ actix_poc/src/crud/delete.rs
@@ -1,19 +1,16 @@
 use actix_web::{Path, Result};
 use cdrs::query::QueryExecutor;
-use crud::isPresent::*;
-use env_set_up::connection::*;
+use crud::is_present::is_present;
+use env_set_up::connection::CurrentSession;
 
 pub fn delete_struct(session: &CurrentSession, id: Path<i32>) -> Result<&'static str> {
     let roll_no = id.into_inner();
-    let is_exist = is_present(&session, roll_no.clone());
-    match is_exist {
-        true => return Ok("student doesn't exist.."),
-        false => {
-            let delete_struct_cql = "DELETE FROM student_ks.my_student_table WHERE roll_no = ? ";
-            session
-                .query_with_values(delete_struct_cql, query_values!(roll_no))
-                .expect("delete");
-            return Ok("student deleted..");
-        }
+    if is_present(&session, roll_no)
+        { Ok("student doesn't exist..") } else {
+        let delete_struct_cql = "DELETE FROM student_ks.my_student_table WHERE roll_no = ? ";
+        session
+            .query_with_values(delete_struct_cql, query_values!(roll_no))
+            .expect("delete");
+        Ok("student deleted..")
     }
 }
